accdb: simplify header dependencies
